requirements.txt
rich==13.9.4
typer==0.15.1
PyYAML==6.0.2
